% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_date.R
\name{check_date_chronology}
\alias{check_date_chronology}
\title{Validate date chronology}
\usage{
check_date_chronology(from_date, to_date)
}
\arguments{
\item{from_date}{Character vector. Start date(s) in \code{"dd/mm/yyyy"} format.}

\item{to_date}{Character vector. End date(s) in \code{"dd/mm/yyyy"} format.}
}
\value{
Logical vector of the same length as the input vectors. \code{TRUE} if
\code{from_date} is chronologically less than or equal to \code{to_date}, otherwise \code{FALSE}.
}
\description{
Checks whether a vector of start dates occurs chronologically before or on a
corresponding vector of end dates.
}
\examples{
check_date_chronology("01/01/2025", "05/01/2025") # TRUE
check_date_chronology("10/01/2025", "01/01/2025") # FALSE
}
