% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nz_holidays.R
\name{download_school_holidays}
\alias{download_school_holidays}
\title{Download School Holiday .ics Files}
\usage{
download_school_holidays(urls, download_dir = tempdir())
}
\arguments{
\item{urls}{Character vector of \code{.ics} file URLs.}

\item{download_dir}{Directory where files should be saved.}
}
\value{
A character vector of saved file paths (invisibly).
}
\description{
Download School Holiday .ics Files
}
