% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecanairquality-package.R
\docType{package}
\name{ecanairquality-package}
\alias{ecanairquality}
\alias{ecanairquality-package}
\title{ecanairquality: Helpful functions to get data and generate plots for ecan air quality data}
\description{
Retrieves data from the Ecan data portal website. Also doe some basic plotting examples.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bws200/ecanairquality}
}

}
\author{
\strong{Maintainer}: Ben Scott \email{ben.scott@ecan.govt.nz}

}
\keyword{internal}
