% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_daily_all_stations.R
\name{get_daily_all_stations}
\alias{get_daily_all_stations}
\title{Retrieve daily air quality data from the ECan website}
\usage{
get_daily_all_stations(
  from_date = "1/01/2026",
  to_date = "31/01/2026",
  on_error = c("ignore", "stop", "warn")
)
}
\arguments{
\item{from_date}{Character. Start date in \code{"dd/mm/yyyy"} format.}

\item{to_date}{Character. End date in \code{"dd/mm/yyyy"} format.}

\item{on_error}{How to handle a station request failure: \code{"ignore"} (the default)
continues silently\verb{,"warn"} continues with other stations and displays a warning
message, \code{"stop"} aborts the request}
}
\value{
A data frame containing daily air quality measurements in
long format.
}
\description{
Retrieves daily averaged air quality data for all monitoring sites
available via the Environment Canterbury (ECan) data portal.
}
