% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nz_holidays.R
\name{get_school_holiday_urls}
\alias{get_school_holiday_urls}
\title{Fetch NZ School Holiday Calendar URLs}
\usage{
get_school_holiday_urls(
  url = "https://www.education.govt.nz/school-terms-and-holidays-dates"
)
}
\arguments{
\item{url}{Base web address to scrape.}
}
\value{
A character vector of absolute URLs.
}
\description{
Scrapes the NZ Ministry of Education website for .ics file download links.
}
