% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nz_holidays.R
\name{get_school_holidays}
\alias{get_school_holidays}
\title{Get NZ School Holiday Dates}
\usage{
get_school_holidays(download_dir = tempdir())
}
\arguments{
\item{download_dir}{Target folder for file downloads. Defaults to a temporary directory.}
}
\value{
A tidy tibble of school holiday dates.
}
\description{
Wraps the fetching, downloading, and parsing steps into a single call.
}
