% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_daily.R
\name{getEcanDaily}
\alias{getEcanDaily}
\title{Retrieve daily data from the Ecan website for all sites}
\usage{
getEcanDaily(from_date, to_date)
}
\arguments{
\item{from_date}{The "from" date to get data.}

\item{to_date}{The "to" date to get data until.}
}
\value{
Daiy data from all sites available
}
\description{
Retrieve daily data from the Ecan website for all sites
}
\examples{
getEcanDaily(from_date = "01/01/2025", to_date = "31/01/2025")
}
