% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nz_holidays.R
\name{parse_school_holidays}
\alias{parse_school_holidays}
\title{Parse and Tidy School Holiday .ics Files}
\usage{
parse_school_holidays(ics_files)
}
\arguments{
\item{ics_files}{Vector of file paths to \code{.ics} files.}
}
\value{
A cleaned tibble of holiday names and dates.
}
\description{
Parse and Tidy School Holiday .ics Files
}
