% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_daily_one_station.R
\name{get_daily_one_station}
\alias{get_daily_one_station}
\title{Title}
\usage{
get_daily_one_station(site_id, from_date, to_date)
}
\arguments{
\item{site_id}{The site ID which can be retrieved by the get_stations function}

\item{from_date}{From date}

\item{to_date}{To date}
}
\description{
Title
}
