% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stations.R
\name{get_stations}
\alias{get_stations}
\title{Title}
\usage{
get_stations()
}
\value{
A data frame containing air quality stations at Environment Canterbury
}
\description{
Title
}
\examples{
\dontrun{
get_stations()
}



}
